requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
casttube==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
googlecontroller==6
idna==3.4
ifaddr==0.2.0
multidict==6.0.4
openai==0.28.1
protobuf==3.20.1
PyChromecast==1.0.3
pydantic==2.4.2
pydantic-core==2.10.1
pyngrok==7.0.0
PyYAML==6.0.1
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.0.6
yarl==1.9.2
zeroconf==0.115.2
